web/requirements.txt
Django>=5.0,<6.0
gunicorn
psycopg2-binary
python-dotenv
whitenoise
Pillow>=10
django-filter>=25.1
requests>=2.32
django-ratelimit>=4.1
djangorestframework>=3.15
django-redis>=5.4
redis>=5
sentry-sdk[django]>=2.0
drf-spectacular[sidecar]
django-cors-headers